CMake: Move _AMD64_ compile def. to x86.cmake
The `_X86_` and `_AMD64_` compile definitions (as well as `_PPC_`)
are used only in the `port/win32/omrsysinfo.c` file. It's interesting,
the definitions are defined only in the `win.cmake` module, and used
only for the win32 port. The `_AMD64_` compile definition should be
moved to the `cmake/modules/platform/arch/x86.cmake` file in line with
other architecture definitions: `_PPC_`, `_PPC64_`, `ARMGNU`, etc.

Issue: #2405
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>