Add the OMRWINDOWS compile definition
As it's discussed with @youngar in the comments to issue #2405,
it would be good to start using macros prefixed with `OMR`.
As more people use OMR, it would be nice if we didn't collide
with commonly used macro names like `WINDOWS`.

The JIT already uses `WINDOWS`, while the rest of the code would
need to be updated.

Constructions such as `#ifdef WIN32` can be replaced by
`#if defined(OMRWINDOWS)`

Constructions such as `#ifdef WIN64` can be replaced by
`#if defined(OMRWINDOWS) && defined(OMR_ENV_DATA64)`

Constructions such as `#if defined(WIN32) || defined(WIN64)`
can be replaced by `#if defined(OMRWINDOWS)`

The `OMRWINDOWS` compile definition is added to the CMake
configuration as well as the autotools one.

Issue: #2405
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>