Replace MS_WINDOWS to OMRWINDOWS in compilertest
It would be good to start using macros prefixed with `OMR`. As more
people use OMR, it would be nice if we didn't collide with commonly
used macro names like WINDOWS, WIN32 or _WIN32.

Some JIT test classes already use MS_WINDOWS. I have no idea where
the macro name is defined but the code is wrapped by the macro looks
like it must be wrapped by the WIN32 or WINDOWS. The commit changes
`MS_WINDOWS` to `OMRWINDOWS` in the `compilertest` module.

Issue: #2405
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>