Remove all occurrences of the WIN32_IBMC macro
`IBMC` is a macro usually defined for `XLC` or the `z/OS` system compiler, and
there isn't a Windows version of `XLC`. Also, the `WIN32_IBMC` macro is defined
neither in CMake nor in autotools. So, these are the reasons to remove
all occurrences of the macro at all.

Issue: #2405
Signed-off-by: Pavel Samolysov <samolisov@gmail.com>