Rename extensible header files in the compiler 1/2
A number of header files in the compiler are designed to be extened by
the consuming runtime. For these headers we prefix the names with OMR
and provide a new default implementation of the header which simply
includes the OMR prefixed header. (this is in line with existing
extensible headers)

Commit is split up into 2 parts in order to help git track the changes
a bit better. First commit renames the files, the second commit creates
stub files (with the old names) which include the new files.

Signed-off-by: Devin Nakamura <devinn@ca.ibm.com>