Implement FindLEX CMake module
Flex is not always available on all platforms which OMR currently
targets. Namely it is not available on z/OS and building it is
non-trivial due to the various dependencies it requires.

Tril is the only consumer of Flex in OMR at the moment. Instead of
continuing with Flex we will start the migration so that the Tril lexer
can be generated with POSIX Lex.

This module is very similar FindFlex.cmake module which is shipped as
part of CMake.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>