Enable GTest <tuple> on z/OS
The C++11 `tuple` class can be included in GTest for z/OS however it is
part of the TR1 extensions, and hence lives in the TR1 namespace. GTest
is configured as such so it expects `tuple` to be in the `std`
namespace. There is a define override (`GTEST_ENV_HAS_STD_TUPLE_`)
however which can be used to import the `tuple` class on z/OS.

We make use of this override along with a code change to enable the use
of system `tuple` class in GTest and it's consumers (Tril).

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>