Define YYLMAX for Lex token limit and clean up xlc compiler directives
`YYLMAX` is a define which is used to override the default token length
limit on z/OS. Because of the copyright headers the token length of a
comment may exceed the default value (100) and Lex will fail to parse
the comment token. We increase the value to 1000 here to avoid such an
issue.

We also take this opportunity to clean up some xlc compile flags which
are not useful or detrimental to compilation of OMR sources on z/OS.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>